Arrays.MultipleStatementAlignment: allow for several equally large groups of arrow alignments
Only applies when `exact` is set to `false`.

If an array has groups of arrows aligned at several different positions, there may be several equally large groups of alignments, i.e.: 3 arrows aligned at position 40 and 3 arrows aligned at position 70.

In that case, either one can be considered for the new non-exact alignment, but one or the other may be smaller than the expected column or larger than the maximum column.

This changes makes sure that all possibilities are considered when determining if there is an existing grouping at an acceptable position to align to.